Move Exception classes to their own namespace (task #9606)
