+++ src/Event/Model/ModelAfterSaveListener.php
@@ -21,8 +21,8 @@ use Cake\Log\LogTrait;
 use Cake\Network\Exception\SocketException;
 use Cake\ORM\Table;
 use Cake\Utility\Inflector;
-use CsvMigrations\UnsupportedForeignKeyException;
-use CsvMigrations\UnsupportedPrimaryKeyException;
+use CsvMigrations\Exception\UnsupportedForeignKeyException;
+use CsvMigrations\Exception\UnsupportedPrimaryKeyException;
 use CsvMigrations\Event\EventName;
 use CsvMigrations\Table as CsvTable;
 use CsvMigrations\Utility\DTZone;
